readme added input validation added

+++ app/src/main/java/com/codesherpa/beerdispenser/app/dtos/BeerDto.java
@@ -1,10 +1,12 @@
 package com.codesherpa.beerdispenser.app.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class BeerDto {
     private Long id; 
     private String name;
-    private Float pricePerLitre;
+    private BigDecimal pricePerLitre;
 }
