+++ app/src/main/java/com/codesherpa/beerdispenser/app/dtos/request/CreateBeerDto.java
@@ -1,5 +1,7 @@
 package com.codesherpa.beerdispenser.app.dtos.request;
 
+import java.math.BigDecimal;
+
 import com.codesherpa.beerdispenser.app.exceptions.ExceptionMessage;
 
 import jakarta.validation.constraints.DecimalMin;
@@ -12,5 +14,5 @@ public class CreateBeerDto {
     public String name;
 
     @DecimalMin(value = "0.0", inclusive = true, message = ExceptionMessage.BEER_PRICE_PER_LITRE_INVALID)
-    public Float pricePerLitre;
+    public BigDecimal pricePerLitre;
 }
